+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/elasticsearch/Queries.scala
@@ -22,10 +22,12 @@ object Queries {
   def publisherQuery(strategy: SearchStrategy)(publisher: FilterValue[String]) = {
     handleFilterValue(publisher, (publisherString: String) =>
       strategy match {
-        case SearchStrategy.MatchAll => matchQuery("publisher.name.keyword_lowercase", publisherString)
+        case SearchStrategy.MatchAll => multiMatchQuery(publisherString)
+          .fields("publisher.acronym", "publisher.name.keyword_lowercase")
+          .minimumShouldMatch("-50%")
         case SearchStrategy.MatchPart =>
           multiMatchQuery(publisherString)
-            .fields("publisher.name", "publisher.name.english")
+            .fields("publisher.acronym", "publisher.name", "publisher.name.english")
             .minimumShouldMatch("-50%")
       }, "publisher.name"
     )